Move utils to XmlParser.java and update comment docs
